Move errors into their own modules
As a developer I would be looking for errors there
