+++ pusher/errors.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+
+class PusherError(Exception):
+    pass
+
+class PusherBadRequest(PusherError):
+    pass
+
+class PusherBadAuth(PusherError):
+    pass
+
+class PusherForbidden(PusherError):
+    pass
+
+class PusherBadStatus(PusherError):
+    pass
